Change the function call chain so that hooking isn't necessary.
Redesign so that UpdateGroups() is called in response to roster or
specialzation updates and it independently calls requests reloading the
layout if the "by-role" layout has changed.  Remove the direct
manipulation of the secure group headers because it can be handled
entirely within GridLayout.

Make the "by-role" layouts make use of more features that can be used in
layouts, as documented in Grid/Layouts.lua.  In particular, use "defaults"
instead of directly setting the same attributes in every group -- using
layout group defaults allows for GridLayout to do the copying for us.