Make role checks more robust against delayed information.
When requesting inspection information from Blizzard using
NotifyInspect(), sometimes INSPECT_READY is not received, and
if it is received, the information we want is not necessarily
available.  Schedule a repeating timer to request notification
for each pending inspection to catch cases where INSPECT_READY
is not seen after NotifyInspect() is called.  Pause the timer
when entering combat and restart it when leaving combat.  Verify
that the information we want is actually available when
INSPECT_READY is received or else keep the inspection request
pending.

Use PLAYER_SPECIALIZATION_CHANGED, which is actually a unit event
and not a player event, to trigger a role check for the unit.
This avoids the dance with UNIT_AURA, which was not terribly
accurate for catching changes in unit specialization.

Queue role checks for units that come back online in case they
disconnected while we were trying to inspect them.

Give pet units a role of "PET" to avoid special checks for pet
units versus non-pet units.  This simplifies the logic in several
places.

Create toggle to allow Blizzard-assigned roles to take precedence.
This toggle defaults to off.