Update roles when the roster updates, not on unit join/left.
This avoids problems with events arriving out of order and ensures
that every member of the group appears in the GridLayoutByRole
layout.

As part of these changes, make the mappings between GUIDs and roles
more DRY by removing the roleGroup table and its corresponding
maintenance code, and simply have the :NameList(role) method
iterate over the roleByGUID table and filter for the specified
role.